WIP: instrument require() for  autobreadcrumbs
