+++ lib/client.js
@@ -9,7 +9,6 @@ var transports = require('./transports');
 var nodeUtil = require('util'); // nodeUtil to avoid confusion with "utils"
 var events = require('events');
 var domain = require('domain');
-var autoBreadcrumbs = require('./breadcrumbs');
 
 var extend = utils.extend;
 
@@ -51,26 +50,9 @@ extend(Raven.prototype, {
 
     // autoBreadcrumbs: true enables all, autoBreadcrumbs: false disables all
     // autoBreadcrumbs: { http: true } enables a single type
-    // this procedure will ensure that this.autoBreadcrumbs is an object populated
-    // with keys -> bools reflecting actual status of all breadcrumb types
-    var autoBreadcrumbDefaults = {
-      console: false,
-      http: false,
-    };
+    this.autoBreadcrumbs = options.autoBreadcrumbs || false;
     // default to 30, don't allow higher than 100
     this.maxBreadcrumbs = Math.max(0, Math.min(options.maxBreadcrumbs || 30, 100));
-    this.autoBreadcrumbs = extend({}, autoBreadcrumbDefaults);
-    if (typeof options.autoBreadcrumbs !== 'undefined') {
-      for (var key in autoBreadcrumbDefaults) {
-        if (autoBreadcrumbDefaults.hasOwnProperty(key)) {
-          if (typeof options.autoBreadcrumbs === 'boolean') {
-            this.autoBreadcrumbs[key] = options.autoBreadcrumbs;
-          } else if (typeof options.autoBreadcrumbs[key] === 'boolean') {
-            this.autoBreadcrumbs[key] = options.autoBreadcrumbs[key];
-          }
-        }
-      }
-    }
 
     this.captureUnhandledRejections = options.captureUnhandledRejections;
     this.loggerName = options.logger || '';
@@ -130,11 +112,7 @@ extend(Raven.prototype, {
       });
     }
 
-    for (var key in this.autoBreadcrumbs) {
-      if (this.autoBreadcrumbs.hasOwnProperty(key)) {
-        this.autoBreadcrumbs[key] && autoBreadcrumbs.instrument(key, this);
-      }
-    }
+    require('./instrumentation/instrumentor')(this, this.autoBreadcrumbs);
 
     this.installed = true;
 
@@ -144,7 +122,7 @@ extend(Raven.prototype, {
   uninstall: function uninstall() {
     if (!this.installed) return this;
 
-    autoBreadcrumbs.restoreOriginals();
+    // autoBreadcrumbs.restoreOriginals();
 
     // todo: this works for tests for now, but isn't what we ultimately want to be doing
     process.removeAllListeners('uncaughtException');
